removed the stupid _priv field
it was used to disallow constructing the struct directly, rather you had
to use the new function. however, that was entirely unnecessary. this is
a library. if you for some reason decide to act un-rusty, only god can
help you anyways, so why should i, a mere mortal, dare to step in your
paths??